+++ scripts/update_redirect_list_TEST.py
